Move file backends to the new structure
