+++ ci/run_tests.sh
@@ -3,7 +3,7 @@
 # - PYTEST_BACKENDS: Space-separated list of backends to run
 # - PYTEST_EXPRESSION: Marker expression, for example "not udf"
 
-TESTS_DIRS="ibis/tests ibis/file/tests"
+TESTS_DIRS="ibis/tests"
 for BACKEND in $PYTEST_BACKENDS; do
     if [[ -d ibis/$BACKEND/tests ]]; then
         TESTS_DIRS="$TESTS_DIRS ibis/$BACKEND/tests"
