+++ ibis/backends/csv/__init__.py
@@ -9,7 +9,7 @@ from ibis.backends.pandas.core import execute, execute_node, pre_execute
 from ibis.backends.pandas.execution.selection import physical_tables
 from ibis.expr.scope import Scope
 from ibis.expr.typing import TimeContext
-from ibis.file.client import FileClient
+from ibis.backends.base_file import FileClient
 
 dialect = PandasDialect
