+++ ibis/backends/hdf5/__init__.py
@@ -3,7 +3,7 @@ import pandas as pd
 import ibis.expr.operations as ops
 import ibis.expr.schema as sch
 from ibis.backends.pandas.core import execute, execute_node
-from ibis.file.client import FileClient
+from ibis.backends.base_file import FileClient
 
 
 def connect(path):
