+++ ibis/backends/parquet/__init__.py
@@ -11,7 +11,7 @@ import ibis.expr.schema as sch
 import ibis.expr.types as ir
 from ibis.backends.pandas import PandasDialect
 from ibis.backends.pandas.core import execute, execute_node
-from ibis.file.client import FileClient
+from ibis.backends.base_file import FileClient
 
 dialect = PandasDialect
