+++ ibis/backends/parquet/tests/conftest.py
@@ -1,0 +1,17 @@
+import pytest
+import pyarrow as pa
+import pyarrow.parquet as pq
+from ibis.backends.parquet import ParquetClient
+
+
+@pytest.fixture
+def parquet(tmpdir, file_backends_data):
+    # create single files
+    d = tmpdir.mkdir('pq')
+
+    for k, v in file_backends_data.items():
+        f = d / '{}.parquet'.format(k)
+        table = pa.Table.from_pandas(v)
+        pq.write_table(table, str(f))
+
+    return ParquetClient(tmpdir).database()
